refactor: isolate run distribution command
